+++ odoo/local-src/connector_esb/models/__init__.py
@@ -20,3 +20,4 @@ from . import res_country
 from . import promotion_alcyon
 from . import product_price
 from . import res_lang
+from . import sale_order_line