+++ HISTORY.rst
@@ -33,6 +33,7 @@ Unreleased
 * ALCN-923: Send an email when a purchase order is canceled
 * ALCN-796: Use the available quantity on product for the reception wizard
 * Create an user with login WSO2 to allow Smile to use restricted controllers
+* ALCN-866: Add webservice for a product a client yearly purchase statistics
 
 **Bugfixes**